+++ doc/whats-new.rst
@@ -7,7 +7,7 @@ What's New
 
 v1.5.0 (unreleased)
 -------------------
-- Added formal support for Python 3.13 (the previous version works fine though)
+- Added formal support for Python 3.13 and 3.14 (the previous version works fine though)
 - Bumped minimum version of psutil from 5.6 to 5.7
 - Handle deprecation in psutil 6